fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int count;
-	int fib1 = 0, fib2 = 1;
+	long fib1 = 0, fib2 = 1;
 	long sum;
 
 	for (count = 0; count < 50; count++)
